app/requirements.txt
GitPython==3.1.31
pygount==1.5.1
pytz==2023.3
tqdm==4.65.0
prettytable==3.6.0